WIP https://github.com/AminAvan/EdgeSimPy/issues/3 . The services possess a 'cpu_cycles_demand', which is a more precise measure of demand compared to simply requesting a core or cores. Therefore, the processing power demand of a service is calculated by the formula (self.cpu_cycles_demand * self.memory_demand), which indicates how much processing power the service requires based on its data and the processes it needs to perform.
On the other hand, on the server side, servers have a 'cpu_cycle' value representing the CPU's clock frequency. The 'processing_power' of the server is then determined by multiplying its 'cpu_cycle' with the number of cores. As a result, the service sends its 'processing power demand' and its deadline (user.delay_slas) to the server. The server then calculates the time it would take to execute the service, which is referred to as 'execution_time'. The server divides 'execution_time' by the deadline (user.delay_slas) to determine the required utilization of its processor. This utilization is a more accurate measure for checking whether the server is overloaded, compared to merely counting the number of cores.